:art: Oppdatert bruk av sidebar
