+++ aksel.nav.no/website/pages/monster-maler/[...slug].tsx
@@ -7,15 +7,15 @@ import { getClient } from "@/sanity/client.server";
 import { getDocuments } from "@/sanity/interface";
 import { destructureBlocks, sidebarQuery } from "@/sanity/queries";
 import {
-  AkselSidebarT,
   AkselTemplatesDocT,
   ArticleListT,
   CodeExampleSchemaT,
   NextPageT,
   ResolveContributorsT,
   ResolveSlugT,
+  SidebarT,
 } from "@/types";
-import { dateStr } from "@/utils";
+import { dateStr, generateSidebar } from "@/utils";
 import { StatusTag } from "@/web/StatusTag";
 import { AkselTable, AkselTableRow } from "@/web/Table";
 import { SEO } from "@/web/seo/SEO";
@@ -26,7 +26,7 @@ import NotFotfund from "../404";
 
 type PageProps = NextPageT<{
   page: ResolveContributorsT<ResolveSlugT<AkselTemplatesDocT>>;
-  sidebar: AkselSidebarT;
+  sidebar: SidebarT;
   seo: any;
   refs: ArticleListT;
   publishDate: string;
@@ -76,7 +76,7 @@ export const getStaticProps: GetStaticProps = async ({
       page,
       slug: slug.slice(0, 2).join("/"),
       seo,
-      sidebar,
+      sidebar: generateSidebar(sidebar, "templates"),
       preview,
       title: page?.heading ?? "",
       id: page?._id ?? "",
